Fix signedness mismatch
NaCl deals with unsigned char buffers, so CArray[uint8] is the correct Raku
type. As elements are moved between these and Buf (which by default is uint8)
we need to use the same types consistently.

This is required for the next Rakudo version which is more picky about int/uint.